requirements.txt
pandas
numpy
prophet
scikit-learn
tensorflow
matplotlib
plotly
streamlit
joblib
python-dotenv
twilio
requests